+++ CHANGELOG.md
@@ -7,10 +7,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- [API] Function: `CelestialBody URT_Registry.GetReceiverCelestialBody(int receiverID)`
+- [API] Function: `CelestialBody URT_Registry.GetTransmitterCelestialBody(int transmitterId)`
+- [API] Interface: `IURT_Transmitter`
+- [API] Interface: `IURT_Receiver`
+- [Gameplay] Occlusion by celestial bodies
+- [Gameplay] Absorption and scattering by atmospheres
+
 ### Changed
 - [API] Refactored core logic to an interface-based model using `IURT_Transmitter` and `IURT_Receiver`.
   - Developer Note: Implementing `PartModules` must now include `KSPFields` named `transmitterID` and `receiverId`. 
   - These fields must remain identical to the `TransmitterID` and `ReceiverId` properties at all times for the solver to function correctly.
+- [Internal] Modified existing and established new custom URT types, deprecating several tuple based collections
+  - `readonly struct URT_BodyValues` (new): Stores cached, squared values for a celestial body
+  - `readonly struct URT_Link` (modified): Now stores an `readonly double AtmosphereAttenuationCoefficient`
+  - `struct URT_ActiveLink` (new): Stores information relevant to a `URT_Link` which is **actively being used**, and which is transient
+  - `readonly struct URT_LinkToProcess` (new): Stores information relevant to a `URT_Link` which is to be evaluated by the network solver. This is transient
+- [Performance] Updated NetworkRebuild coroutine to run on realtime delays, not Unity timescale, which reduces performance impact
+- [Performance] Network solver uses Branch and Bound logic to guarantee maximum efficiency while avoiding excessive link occlusion evaluations.
+### Fixed
+- Coroutines only running once in `URT_Registry`
 
 ## 0.0.1 - 2026-06-05